OpenEIT/dashboard/requirements.txt
pyserial==3.4
numpy==1.14.5
scipy==1.1.0
plotly==3.0.2
dash==0.21.1
dash_renderer==0.13.0
dash_core_components==0.25.1
dash_html_components==0.11.0
python-slugify==1.2.6
